VUE: Move vue file which is not belong to view dir
